+++ Twinder/DeepLinks.swift
@@ -7,12 +7,24 @@
 //
 
 import Foundation
+import UIKit
 
 func twtrScreenName(screen_name: String){
-    let deepLinkUrl: NSString = "twitter://user?screen_name=" + screen_name
-    let url = NSURL(string: (deepLinkUrl))
-    if UIApplication.sharedApplication().canOpenURL(url!) {
-        // url is good to go
-        UIApplication.sharedApplication().openURL(url!)
+    let allowedScreenNameCharacters = NSCharacterSet(charactersInString: "ABCDEFGHIJKLMNOPQRSTUVWXYZabcdefghijklmnopqrstuvwxyz0123456789_")
+    let screenNameLength = (screen_name as NSString).length
+    if screenNameLength == 0 || screenNameLength > 15 ||
+        screen_name.rangeOfCharacterFromSet(allowedScreenNameCharacters.invertedSet) != nil {
+        return
+    }
+
+    let components = NSURLComponents()
+    components.scheme = "twitter"
+    components.host = "user"
+    components.queryItems = [NSURLQueryItem(name: "screen_name", value: screen_name)]
+
+    if let url = components.URL {
+        if UIApplication.sharedApplication().canOpenURL(url) {
+            UIApplication.sharedApplication().openURL(url)
+        }
     }
 }